Extract component hook
