+++ GameDocumentEngine.Ui/src/doc-types/Dashboard/info/InfoWidgetModal.tsx
@@ -2,17 +2,11 @@ import { Button } from '@/components/button/button';
 import { Prose } from '@/components/text/common';
 import type { ModalContentsProps } from '@/utils/modal/modal-service';
 import type { Widget } from '../types';
-import { NamedIcon } from '@/components/named-icon/NamedIcon';
+import { useDocumentName } from '@/components/named-icon/useDocumentName';
 import { ModalDialogLayout } from '@/utils/modal/modal-dialog';
 import type { WidgetBase } from '@/documents/defineDocument';
 import { WidgetContainer } from '../grid-utils';
-import {
-	useDocTypeTranslation,
-	useDocument,
-	useDocumentType,
-	useTranslationFor,
-} from '@/utils/api/hooks';
-import { missingDocumentType } from '@/documents/defaultMissingWidgetDefinition';
+import { useDocTypeTranslation, useTranslationFor } from '@/utils/api/hooks';
 import { useWidget } from '../useWidget';
 
 export function InfoWidgetModal<TWidget extends WidgetBase>({
@@ -25,26 +19,19 @@ export function InfoWidgetModal<TWidget extends WidgetBase>({
 		widget: Widget<TWidget>;
 	}
 >) {
-	const document = useDocument(gameId, widget.documentId);
-	const { icon: Icon } =
-		useDocumentType(gameId, widget.documentId)?.typeInfo ?? missingDocumentType;
+	const DocumentName = useDocumentName(gameId, widget.documentId);
 	const Widget = useWidget(gameId, widget);
-	const t = useDocTypeTranslation('Dashboard');
-	const tDocument = useTranslationFor(gameId, widget.documentId);
+	const t = useDocTypeTranslation('Dashboard', {
+		keyPrefix: 'info-widget-modal',
+	});
 	const tWidget = useTranslationFor(gameId, widget.documentId, widget.widget);
 	// TODO: better layout
 	return (
 		<ModalDialogLayout>
-			<ModalDialogLayout.Title>
-				{t('info-widget-modal.title')}
-			</ModalDialogLayout.Title>
+			<ModalDialogLayout.Title>{t('title')}</ModalDialogLayout.Title>
 			<div className="flex flex-row-reverse flex-wrap justify-end gap-4">
 				<Prose>
-					<NamedIcon
-						name={document.name}
-						icon={Icon}
-						typeName={tDocument('name')}
-					/>
+					<DocumentName />
 					<br />
 					{tWidget('name')}
 				</Prose>
@@ -53,7 +40,7 @@ export function InfoWidgetModal<TWidget extends WidgetBase>({
 				</WidgetContainer>
 			</div>
 			<ModalDialogLayout.Buttons>
-				<Button onClick={() => resolve()}>{t('info-widget-modal.ok')}</Button>
+				<Button onClick={() => resolve()}>{t('ok')}</Button>
 			</ModalDialogLayout.Buttons>
 		</ModalDialogLayout>
 	);
